move @log into its own module
